add init --jsonnet option
